QuestCatalog/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
bcrypt==4.2.0
click==8.1.7
Django==4.2.15
djangorestframework==3.15.2
exceptiongroup==1.2.2
fastapi==0.112.0
fastapi-jwt-auth==0.5.0
functions==0.7.0
h11==0.14.0
idna==3.7
iniconfig==2.0.0
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
pydantic==1.10.7
pydantic_core==2.20.1
PyJWT==1.7.1
PyMySQL==1.1.1
pytest==8.3.2
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.32
sqlparse==0.5.1
starlette==0.37.2
tomli==2.0.1
typing_extensions==4.12.2
uvicorn==0.30.5
